RFC: rewrite inifix's CLIs as a typer app
