+++ .github/workflows/ci.yml
@@ -24,6 +24,9 @@ jobs:
         - '3.11'
         - '3.12'
         - '3.13'
+        sync-options:
+        - --extra cli
+        - ''
         include:
         - os: macos-latest
           python-version: 3.10.0
@@ -51,7 +54,7 @@ jobs:
     - name: Run tests (with coverage)
       if: matrix.os == 'ubuntu-latest'
       run: |
-        uv run --frozen --no-editable --group covcheck \
+        uv run --frozen --no-editable --group covcheck ${{ matrix.sync-options }} \
           coverage run --parallel-mode -m pytest --color=yes --doctest-modules
 
     - name: Upload coverage data
@@ -60,7 +63,7 @@ jobs:
       if: matrix.os == 'ubuntu-latest'
       uses: actions/upload-artifact@6f51ac03b9356f520e9adb1b1b7802705f340c2b # v4.5.0
       with:
-        name: inifix_coverage_data-${{ matrix.os }}-${{ matrix.python-version }}
+        name: inifix_coverage_data-${{ matrix.os }}-${{ matrix.python-version }}-${{ matrix.sync-options }}
         path: .coverage.*
         if-no-files-found: ignore
         include-hidden-files: true
@@ -84,7 +87,7 @@ jobs:
     - run: uv sync --frozen --no-editable --group concurrency
     - run: |
         uv run --no-sync \
-          pytest --color=yes --count 100 tests/test_concurrent.py
+          pytest --color=yes --count 100 -k concurrency
 
   coverage:
     name: Combine & check coverage
@@ -141,7 +144,7 @@ jobs:
       if: matrix.typechecker == 'pyright'
       run: |
         uv run --frozen --no-editable --group typecheck \
-          pyright --verifytypes inifix
+          pyright --ignoreexternal --verifytypes inifix
 
   optimized_python:
     runs-on: ubuntu-latest
