+++ .pre-commit-hooks.yaml
@@ -1,12 +1,16 @@
 - id: inifix-format
   name: Format .ini files
   language: python
-  entry: inifix-format
-  description: Format configuration files with inifix-format
+  entry: inifix format
+  description: Format configuration files with inifix format
   files: (\.ini)$
+  # keep in sync with pyproject.toml
+  additional_dependencies: [typer-slim>=0.15.1, rich>=13.9.4]
 - id: inifix-validate
   name: Validate .ini files (no formatting)
   language: python
-  entry: inifix-validate
-  description: Validate configuration files with inifix-validate
+  entry: inifix validate
+  description: Validate configuration files with inifix validate
   files: (\.ini)$
+  # keep in sync with pyproject.toml
+  additional_dependencies: [typer-slim>=0.15.1, rich>=13.9.4]
