+++ pyproject.toml
@@ -27,12 +27,18 @@ file = "README.md"
 content-type = "text/markdown"
 
 [project.scripts]
-inifix-format = "inifix.format:main"
-inifix-validate = "inifix.validate:main"
+inifix = "inifix.__main__:app"
 
 [project.urls]
 Homepage = "https://github.com/neutrinoceros/inifix"
 Changelog = "https://github.com/neutrinoceros/inifix/blob/main/CHANGELOG.md"
+
+[project.optional-dependencies]
+# keep in sync with .pre-commit-hooks.yaml
+cli = [
+    "rich>=13.9.4",
+    "typer>=0.15.1",
+]
 
 [dependency-groups]
 test = [
@@ -40,6 +46,7 @@ test = [
     "pytest>=6.1",
 ]
 concurrency = [
+    "inifix[cli]",
     {include-group = "test"},
     "pytest-repeat>=0.9.3",
 ]
@@ -49,6 +56,7 @@ covcheck = [
     "coverage>=6.5.0 ; python_version >= '3.11'",
 ]
 typecheck = [
+    "inifix[cli]",
     "mypy>=1.11.2",
     "pyright>=1.1.390",
 ]
