+++ src/inifix/__init__.py
@@ -2,8 +2,8 @@ from .io import dump
 from .io import dumps
 from .io import load
 from .io import loads
-from .validation import validate_inifile_schema
-from .format import format_string
+from ._validation import validate_inifile_schema
+from ._format import format_string
 from ._version import *
 
 __all__ = [
